[SonDTT] Thêm xử lý thuật toán FB

+++ Tuan 2/BT_1A/14_1A/14_1A/Program.cs
@@ -19,13 +19,16 @@ namespace _14_1A
             string input = args[0];
             string outputFI = args[1];
             string outputAR = args[2];
-            double minsup = double.Parse(args[3]);
-            double minconf = double.Parse(args[4]);
+            float minsup = float.Parse(args[3]);
+            float minconf = float.Parse(args[4]);
 
             List<Column> InputData = MyFunction.readCSV(input);
             MyFunction.printColumnList(InputData);
 
-
+			//List<Row> InputData = MyFunction.readCSV2(input);
+            //FPGrowth fpGrowth = new FPGrowth();
+            //fpGrowth.CreateFPTreeAndGenerateFrequentItemsets(InputData, minsup);
+            
             Console.WriteLine("Press any key to exit...");
             Console.ReadKey();
             return;
